solve err in comunication db

+++ src/domain/user.rs
@@ -1,6 +1,6 @@
 
 pub struct User {
-    pub id: u64,
+    pub id: i32,
     pub username: String,
     pub password: String,
     pub email: String,
